# Only create an array with default options if we have default options